+++ lib/arachni/element/ui_form/dom.rb
@@ -55,6 +55,10 @@ class DOM < DOM
         @valid_input_names.include? name.to_s
     end
 
+    def coverage_id
+        "#{super}:#{@method}#{locator.hash}"
+    end
+
     def type
         self.class.type
     end